Using unique() method from the Eloquent\Collection class
Refer to https://github.com/laravel/framework/issues/24727 for more information, but as a summary:

There are two Collection classes in Laravel:
Illuminate\Support\Collection and Illuminate\Database\Eloquent\Collection

The first one is the base class for the second one. When you fetch models from the database, you receive a collection of the second class. Certain methods (e.g. pluck()) convert this collection to the base class.

The second class uses the models' primary key to remove duplicates (the behavior you are expecting).
The first class, however, compares the actual values.